% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_parameter_skew}
\alias{apply_parameter_skew}
\title{Pivot the parameters passed into the function so relationships between the
parameters remain consistent but giving control to providing more or less
focus on extreme bins}
\usage{
apply_parameter_skew(params, skew)
}
\arguments{
\item{params}{numeric vector of parameters; assumed to be in order of
increasing waiting times (bins)}

\item{skew}{numeric; length 1, a multiplier to be used on the final
parameter. A skew of 1 will keep the params identical to the input params.
Value must be greater than 0}
}
\value{
a revised numeric vector of parameters
}
\description{
Pivot the parameters passed into the function so relationships between the
parameters remain consistent but giving control to providing more or less
focus on extreme bins
}
\details{
The skew parameter is applied to the final item of the params
  vector. The inverse of the skew parameter is applied to the second item of
  the params vector. The adjustment made to the items in between item to and
  item n are calculated by a linear extrapolation between the inverse skew
  and the skew. The first item is held constant.
}
\examples{
apply_parameter_skew(
  params = c(0.03, 0.02, 0.02, 0.01, 0.04, 0.05),
  skew = 1.05
)
}
