% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{apply_params_to_projections}
\alias{apply_params_to_projections}
\title{Apply the months waited parameters for renege and capacity to projections and
capacity and referrals If needed, or if validating your parameters, include
projections for incomplete pathways per period}
\usage{
apply_params_to_projections(
  capacity_projections,
  referrals_projections,
  incomplete_pathways = NULL,
  regen_capacity_params
)
}
\arguments{
\item{capacity_projections}{numeric; vector of projections for capacity for
each time step. This must be the same length as referrals_projections}

\item{referrals_projections}{numeric; vector of projections for reneges for
each time step. This must be the same length as capacity_projections}

\item{incomplete_pathways}{tibble; two column data frame or tibble, with
fields called months_waited_id (taking values 0 to the maximum months
waited group of interest), and incompletes (the count of the number of
incomplete pathways) representing the count of incomplete pathways at
timestep 0 (to initialise the model with)}

\item{regen_capacity_params}{tibble; three column data frame or tibble, with
fields called months_waited_id (taking values 0 to the maximum months
waited group of interest), and fields called cap_param and renege_param,
which are outputs from the function
\code{calibrate_capacity_renege_params()}}
}
\description{
Apply the months waited parameters for renege and capacity to projections and
capacity and referrals If needed, or if validating your parameters, include
projections for incomplete pathways per period
}
