% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{calc_wl_sizes}
\alias{calc_wl_sizes}
\title{Calculate Waiting List Sizes by Time Waited}
\usage{
calc_wl_sizes(wl_removals, referrals)
}
\arguments{
\item{wl_removals}{A data frame containing the renege rate and potential number of
services per waiting compartment. It must include columns \code{r} (removal/renege
rate), and \code{service} (number of potential services).}

\item{referrals}{numeric of length 1; value representing the steady-state number
of referrals.}
}
\value{
A data frame with calculated \code{wlsize} and \code{sigma} columns for each row.
}
\description{
This function applies the available capacity, defined by the geometric
distribution and the total capacity, and updates a data frame representing
the waiting list, by time waited, by calculating the waiting list size
(\code{wlsize}) and the number of serviced items (\code{sigma}) for each time
compartment waited.
}
\details{
For each row in \code{wl_removals}, the function computes the new waiting
  list size based on the previous size, the removal rate, and the number
  serviced. If the computed size is positive, it is assigned to \code{wlsize} and
  \code{sigma} is set to the serviced value. If not, \code{wlsize} is set to zero and
  \code{sigma} is set to the floored value of the previous size after removal.
}
\examples{
\dontrun{
wl_removals <- data.frame(r = c(0.1, 0.2), service = c(5, 3))
calc_wl_sizes(wl_removals, referrals = 50)
}
}
