% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_incompletes}
\alias{calculate_incompletes}
\title{Calculate the number of incomplete pathways in a bin from the inflow, reneges
and treatments. Following the calculations, apply an optional redistribution
of the occasions where incompletes are negative (eg, more treatments have
been performed than there were people).}
\usage{
calculate_incompletes(inflow, reneges, treatments, redistribution_method)
}
\arguments{
\item{inflow}{numeric; vector of count of patients moving into a bin in a
period in order of lowest to highest bin}

\item{reneges}{numeric; vector of count of patients reneging from a bin in a
period in order of lowest to highest bin}

\item{treatments}{numeric; vector of count of patients being treated from a
bin in a period in order of lowest to highest bin}

\item{redistribution_method}{string; one of "none", "evenly" or
"prioritise_long_waiters"}
}
\description{
Calculate the number of incomplete pathways in a bin from the inflow, reneges
and treatments. Following the calculations, apply an optional redistribution
of the occasions where incompletes are negative (eg, more treatments have
been performed than there were people).
}
