% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_timestep_transitions}
\alias{calculate_timestep_transitions}
\title{Calculates the flow from each stock at each timestep}
\usage{
calculate_timestep_transitions(data, max_months_waited)
}
\arguments{
\item{data}{data frame with fields called months_waited_id, type, period_id,
value. Type can take the values "Referrals", "Incomplete" and "Complete".
Referrals will only have "months_waited_id" = 0. period_id is an integer
representing the chronology of the data.}

\item{max_months_waited}{integer; the maximum number of months to group
patients waiting times by for the analysis. Data are published up to 104
weeks, so 24 is likely to be the maximum useful value for this argument.}
}
\description{
Calculates the flow from each stock at each timestep
}
