% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_months_waited_to_id}
\alias{convert_months_waited_to_id}
\title{convert the string version of months waited to the numeric id version}
\usage{
convert_months_waited_to_id(months_waited, max_months_waited)
}
\arguments{
\item{months_waited}{string; vector with format, by example "2-3"}

\item{max_months_waited}{integer; the maximum number of months to group
patients waiting times by for the analysis. Data are published up to 104
weeks, so 24 is likely to be the maximum useful value for this argument.}
}
\description{
convert the string version of months waited to the numeric id version
}
\examples{

mnths_waited <- c("<1", "1-2", "2-3", "3-4", "4-5", "5+")
convert_months_waited_to_id(
  months_waited = mnths_waited,
  max_months_waited = 3
)
}
