% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_dummy_data}
\alias{create_dummy_data}
\title{Create a set of dummy data to put through the functions of the package}
\usage{
create_dummy_data(
  type,
  max_months_waited,
  number_periods,
  referral_values = 500:700,
  max_incompletes = 500,
  max_treatments = 500,
  seed = 123
)
}
\arguments{
\item{type}{string; one of "referrals", "incompletes" or "completes"}

\item{max_months_waited}{integer; the maximum number of months to group
patients waiting times by for the analysis. Data are published up to 104
weeks, so 24 is likely to be the maximum useful value for this argument.}

\item{number_periods}{integer; the intended number of periods in the dataset}

\item{referral_values}{integer; vector of values that are sampled from for
the count of referrals at each time step}

\item{max_incompletes}{integer; the maximum number of incomplete pathways
possible in one time step}

\item{max_treatments}{integer; the maximum number of treatments possible}

\item{seed}{seed to generate the random data from}
}
\value{
a tibble whose columns depend on the type input. If type is
  "referrals" then it will have two fields, period_id and referrals. If type
  is "completes" or "incompletes", the fields will be period_id,
  months_waited_id and treatments/incompletes, depending on the type value
}
\description{
Create a set of dummy data to put through the functions of the package
}
\examples{
create_dummy_data(
  type = "referrals",
  max_months_waited = 4,
  number_period = 6
)
}
