% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_temp_file}
\alias{download_temp_file}
\title{Download a file from a url to a temporary file}
\usage{
download_temp_file(excel_url)
}
\arguments{
\item{excel_url}{sting; url of the file. Must have an xlsx or xls extension}
}
\value{
filepath to where the temporary file is stored
}
\description{
Download a file from a url to a temporary file
}
