% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{find_p}
\alias{find_p}
\title{Find the value of p1 that achieves a target waiting time percentile}
\usage{
find_p(
  target_time = 18/4.35,
  renege_params,
  mu_1,
  p1_lower = 0.1,
  p1_upper = 0.85,
  tolerance = 0.001,
  referrals,
  percentile = 0.92
)
}
\arguments{
\item{target_time}{numeric of length 1; the target waiting time,
in weeks, to achieve the specified percentile.}

\item{mu_1}{numeric of length 1; treatment capacity in the first month.}

\item{p1_lower}{numeric of length 1; lower bound for the binary search of \code{p1}.
Must be between 0 and 1 and less than  \code{p1_upper}. Defaults to 0.1.}

\item{p1_upper}{numeric of length 1; upper bound for the binary search of \code{p1}.
Must be between 0 and 1 and more than  \code{p1_lower}. Defaults to 0.85.}

\item{tolerance}{numeric of length 1; tolerance for the binary search stopping
criterion. The unit is weeks.}

\item{referrals}{numeric of length 1; value representing the steady-state number
of referrals.}

\item{percentile}{Numeric value between 0 and 1 indicating the desired percentile (default is 0.92).}
}
\value{
A list containing: \item{p1}{The value of \code{p1} found to achieve
  the target waiting time.} \item{time_p}{The interpolated waiting time at
  the specified percentile.} \item{df}{The data frame with historical data
  formatted and waiting list sizes calculated.} \item{niterations}{The number
  of iterations performed in the binary search.}
}
\description{
Uses a binary search algorithm to find the value of \code{p1} (the proportion
parameter) that achieves a specified target waiting time (\code{target_time})
at a given percentile (hardcoded as 92nd percentile) of the cumulative
waiting list size, given a treatment capacity in the first month
(\code{mu_1}) and other parameters.
}
\details{
The function repeatedly uses the removals table (from \code{initialise_removals()})
  and calculates waiting list sizes for candidate values of \code{p1}, using binary search to
  home in on the value that achieves the target waiting time at the required
  percentile. The search stops when the difference between upper and lower
  bounds is less than \code{tolerance}, or when the calculated waiting time is
  within \code{tolerance} of the target.
}
\seealso{
\code{\link{initialise_removals}}, \code{\link{calc_wl_sizes}},
  \code{\link{hist_percentile_calc}}
}
