% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{geom_fn_s}
\alias{geom_fn_s}
\title{Calculate Geometric Survival Probabilities}
\usage{
geom_fn_s(p1, max_num_months = 24)
}
\arguments{
\item{p1}{numeric of length 1; the probability parameter for the geometric distribution
(probability of failure in each period).}

\item{max_num_months}{integer of length 1; the maximum number of months to calculate
survival probabilities for, between 1 and 24. Default is 24.}
}
\value{
A numeric vector of survival probabilities for each month.
}
\description{
Computes the survival probabilities for a geometric distribution over a
specified number of months.
}
\examples{
\dontrun{
# Calculate the geometric survival probabilities for p1 = 0.2, up to 24 months
geom_fn_s(p1 = 0.2, max_num_months = 24)

# Calculate for a single month
geom_fn_s(p1 = 0.7, max_num_months = 1)
}

}
