% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_rtt_data}
\alias{get_rtt_data}
\title{Download and tidy the referral to treatment data from the NHS Statistics
webpage}
\usage{
get_rtt_data(
  type,
  url = "https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/",
  date_start = as.Date("2019-04-01"),
  date_end = Sys.Date(),
  show_progress = FALSE
)
}
\arguments{
\item{type}{string; one of "complete", "incomplete" or "referral"}

\item{url}{string; url of the NHS Referral to Treatment (RTT) Waiting Times}

\item{date_start}{date; start date (earliest date is 1st April 2016, but the
default is 1st April 2019)}

\item{date_end}{date; end date (defaults to "today")}

\item{show_progress}{logical; show progress of downloading and processing
files. Defaults to false}
}
\value{
a tibble with fields for trust, specialty, period, type,
  months_waited and value
}
\description{
Download and tidy the referral to treatment data from the NHS Statistics
webpage
}
\examples{
\dontrun{
monthly_rtt <- NHSRtt::get_rtt_data(
  type = "complete",
  date_start = as.Date("2024-10-01"),
  date_end = as.Date("2024-11-01"),
  show_progress = TRUE
)
}

}
