% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{hist_percentile_calc}
\alias{hist_percentile_calc}
\title{Calculate the Time at a Given Percentile of a Weighted Distribution}
\usage{
hist_percentile_calc(
  wl_structure,
  percentile = 0.92,
  wlsize_col = "wlsize",
  time_col = "months_waited_id"
)
}
\arguments{
\item{wl_structure}{A data frame containing at least the columns specified by \code{wlsize_col} and \code{time_col}.}

\item{percentile}{Numeric value between 0 and 1 indicating the desired percentile (default is 0.92).}

\item{wlsize_col}{Character string specifying the name of the column representing the weights (default is \code{wlsize}).}

\item{time_col}{Character string specifying the name of the column representing the time variable (default is \code{time}).}
}
\value{
A numeric value representing the interpolated time at which the cumulative sum of the weights reaches the specified
  percentile. The value is the number of months waited where the percentile value is met. Despite the requirement that the
  \code{time_col} starts at 0, a value of 0.5 means the percentile value is met after waiting 0.5 months
}
\description{
Computes the interpolated time at which a specified percentile of the cumulative sum
of a weight column (e.g., waiting list size) is reached, based on a sorted data frame.
}
\details{
The function sorts the data frame by the time column, computes the cumulative sum of the weights,
and determines the time at which the cumulative sum first meets or exceeds the specified percentile
of the total sum. If the percentile falls within a bin, linear interpolation is used to estimate the precise time.
}
\examples{
\dontrun{
wl_structure <- data.frame(time = 1:10, wlsize = c(5, 3, 8, 2, 7, 4, 6, 1, 9, 2))
hist_percentile_calc(wl_structure, percentile = 0.9)
}
}
