% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{identify_n_skip_rows}
\alias{identify_n_skip_rows}
\title{Find the first row of the table to import}
\usage{
identify_n_skip_rows(filepath, sheet = "Provider")
}
\arguments{
\item{filepath}{file path for the excel file}

\item{sheet}{string; sheet name}
}
\value{
integer representing the number of rows to skip in the sheet until
  the header column of interest
}
\description{
Find the first row of the table to import
}
