% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{initialise_removals}
\alias{initialise_removals}
\title{Configure Removals Data Frame}
\usage{
initialise_removals(renege_params, p1, mu)
}
\arguments{
\item{renege_params}{numeric; proportion \ifelse{html}{\out{(0 &le; r<sub>m</sub> &leq; 1)}}{\eqn{(0 \leq r_m \leq 1)}}
of waiting patients reneging in the m-th month of waiting, where \ifelse{html}{\out{r<sub>2</sub>}}{\eqn{r_2}} = 0.1
refers to 10\% of those in the second month of waiting reneging.}

\item{p1}{numeric of length 1; the probability parameter for the geometric distribution
(probability of failure in each period).}

\item{mu}{numeric of length 1; total capacity to be applied over the geometric
distribution}
}
\value{
A data frame with columns: \code{months_waited_id}, \code{r}, and
  \code{service}. \code{months_waited_id} is the lower bound of the
  number of months waited. \code{r} is the renege rate for each month
  waited compartment. \code{service} is the initial capacity provided,
  applied across the compartments using the geometric distribution
  defined by \code{p1}.
}
\description{
Constructs a data frame representing the configuration of removals,
  including time, renge rates, and potential services.
}
\examples{
# Example usage:
set.seed(123)
renege_params <- runif(24)
initialise_removals(renege_params, p1 = 0.5, mu = 50)

}
\seealso{
\code{\link{geom_fn_s}}
}
