% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{latest_orgs}
\alias{latest_orgs}
\title{Returns a table of organisations and how they map to one another from the
latest available RTT file}
\usage{
latest_orgs(
  url = "https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/"
)
}
\arguments{
\item{url}{string; url of the NHS Referral to Treatment (RTT) Waiting Times}
}
\value{
a tibble with fields for "NHS Region Code", "NHS Region Name",
  "Provider", "Parent Org Code", "Provider Parent Name", "Provider Org Code",
  "Provider Org Name", "Commissioner", "Parent Org Code", "Commissioner
  Parent Name", "Commissioner Org Code",	"Commissioner Org Name"
}
\description{
Returns a table of organisations and how they map to one another from the
latest available RTT file
}
