% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{latest_rtt_date}
\alias{latest_rtt_date}
\title{Returns the date of the latest available data}
\usage{
latest_rtt_date(
  url = "https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/"
)
}
\arguments{
\item{url}{string; url of the NHS Referral to Treatment (RTT) Waiting Times}
}
\value{
the latest date of available data
}
\description{
Returns the date of the latest available data
}
\examples{
latest_rtt_date()
}
