% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{latest_rtt_file}
\alias{latest_rtt_file}
\title{Returns the url for the latest available file}
\usage{
latest_rtt_file(
  url = "https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/"
)
}
\arguments{
\item{url}{string; url of the NHS Referral to Treatment (RTT) Waiting Times}
}
\description{
Returns the url for the latest available file
}
