% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obtain_links}
\alias{obtain_links}
\title{Identify links within a given url}
\usage{
obtain_links(url)
}
\arguments{
\item{url}{string; url of interest to identify the urls within}
}
\value{
named vector with urls linked to within the provided url
}
\description{
Identify links within a given url
}
