% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{optimise_steady_state_lp}
\alias{optimise_steady_state_lp}
\title{Linear programming solution to steady state optimisation}
\usage{
optimise_steady_state_lp(
  renege_params,
  referrals,
  gamma,
  percentile,
  theta,
  s_given
)
}
\arguments{
\item{renege_params}{numeric; proportion \ifelse{html}{\out{(0 &le; r<sub>m</sub> &leq; 1)}}{\eqn{(0 \leq r_m \leq 1)}}
of waiting patients reneging in the m-th month of waiting, where \ifelse{html}{\out{r<sub>2</sub>}}{\eqn{r_2}} = 0.1
refers to 10\% of those in the second month of waiting reneging.}

\item{referrals}{numeric of length 1; value representing the steady-state number
of referrals.}

\item{gamma}{numeric vector, length equal to \code{length(renege_params)}; must take values between
0 and 1. Vector represents the proportion of the corresponding compartment that the `percentile`
target refers. For example, if the data has five compartments and the target percentile was at
exactly 3.5 months, the corresponding vector would be c(1, 1, 1, 0.5, 0, 0)}

\item{percentile}{numeric value between 0 and 1 indicating the desired percentile (default is 0.92).}

\item{theta}{numeric value of length 1, must be a value between 0 and 1 which is equal to
the proportion of all departures that are reneges}

\item{s_given}{numeric vector, length equal to \code{length(renege_params)}; values must be between
0 and 1 and must represent the proportion of total treatments that are applied to each compartment.
This is used as a constraint in the optimisation and the output should follow the same profile}
}
\description{
Linear programming solution to steady state optimisation
}
