% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{redistribute_incompletes}
\alias{redistribute_incompletes}
\title{Redistribute the counts of incompletes within a vector where they are
negative into the positive values}
\usage{
redistribute_incompletes(incomplete_counts)
}
\arguments{
\item{incomplete_counts}{numeric; vector of incomplete counts}
}
\description{
Redistribute the counts of incompletes within a vector where they are
negative into the positive values
}
