% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{redistribute_incompletes_optimally}
\alias{redistribute_incompletes_optimally}
\title{Redistribute cases where there are more treatments and reneges than people in
a particular bin for a particular period. This function chooses the
redistribute the surplus treatment to the people waiting the longest first}
\usage{
redistribute_incompletes_optimally(incomplete_counts)
}
\arguments{
\item{incomplete_counts}{numeric; vector of count of patients with an incomplete
pathway within a bin in a period in order of lowest to highest bin}
}
\description{
Redistribute cases where there are more treatments and reneges than people in
a particular bin for a particular period. This function chooses the
redistribute the surplus treatment to the people waiting the longest first
}
