% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtt_links.R
\name{rtt_links_by_year}
\alias{rtt_links_by_year}
\title{Get RTT page links for a range of years}
\usage{
rtt_links_by_year(url, start_month, end_month)
}
\arguments{
\item{url}{The base RTT stats webpage address.}

\item{start_month}{The first month in the range of links wanted.}

\item{end_month}{The last month in the range of links wanted.}
}
\value{
A character vector of webpage
}
\description{
Generates links by financial year to the RTT statistics webpages. The
financial year runs from April to March. So for example from "2019-01-01" to
"2024-01-01" should create links for 2018-19 to 2023-24.
}
\details{
Appends to the base url in the format rtt-data-yyyy-yy
}
