% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substr_right}
\alias{substr_right}
\title{Subset string from the right}
\usage{
substr_right(x, n)
}
\arguments{
\item{x}{the string}

\item{n}{integer; number of letters to subset from on the right}
}
\value{
subsetted string
}
\description{
Subset string from the right
}
