% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{calibrate_capacity_renege_params}
\alias{calibrate_capacity_renege_params}
\title{calculate the capacity and renege parameters over the period of data per
month waited}
\usage{
calibrate_capacity_renege_params(data, max_months_waited = 12)
}
\arguments{
\item{max_months_waited}{integer; the maximum number of months to group
patients waiting times by for the analysis. Data are published up to 104
weeks, so 24 is likely to be the maximum useful value for this argument.}
}
\description{
calculate the capacity and renege parameters over the period of data per
month waited
}
