% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_rtt_data}
\alias{get_rtt_data}
\title{Download and tidy the referral to treatment data from the NHS Statistics
webpage}
\usage{
get_rtt_data(
  url = "https://www.england.nhs.uk/statistics/statistical-work-areas/rtt-waiting-times/",
  date_start = as.Date("2019-04-01"),
  date_end = Sys.Date(),
  show_progress = FALSE
)
}
\arguments{
\item{url}{string; url of the NHS Referral to Treatment (RTT) Waiting Times}

\item{date_start}{date; start date (earliest date is 1st April 2011, but the
default is 1st April 2019)}

\item{show_progress}{logical; show progress of downloading and processing
files. Defaults to false}

\item{date_end;}{date; end date (defaults to "today")}
}
\value{
a tibble with fields for trust, specialty, period, type,
  months_waited and value
}
\description{
Download and tidy the referral to treatment data from the NHS Statistics
webpage
}
