% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimiser.R
\name{optimise_capacity}
\alias{optimise_capacity}
\title{Optimise the capacity profile for projections}
\usage{
optimise_capacity(
  t_1_capacity,
  referrals_projections,
  incomplete_pathways,
  renege_capacity_params,
  max_months_waited,
  target,
  tolerance
)
}
\arguments{
\item{t_1_capacity}{numeric; a single value for the capacity for the first
time period}

\item{referrals_projections}{numeric; a vector for the number of referrals
for each period in the projected time period}

\item{incomplete_pathways}{tibble; two column data frame or tibble, with
fields called months_waited_id (taking values 0 to the maximum months
waited group of interest), and incompletes (the count of the number of
incomplete pathways) representing the count of incomplete pathways at
timestep 0 (to initialise the model with)}

\item{renege_capacity_params}{tibble; three column data frame or tibble, with
fields called months_waited_id (taking values 0 to the maximum months
waited group of interest), and fields called capacity_param and
renege_param, which are outputs from the function
\code{calibrate_capacity_renege_params()}}

\item{max_months_waited}{integer; the maximum number of months to group
patients waiting times by for the analysis. Data are published up to 104
weeks, so 24 is likely to be the maximum useful value for this argument.}

\item{target}{string length 1; can be either a percentage point change, eg,
"~-5%" or a percent value, eg, "5%"}

\item{tolerance}{numeric length 1; the tolerance used to compare the absolute
error with in the max_months_waited bin to determine convergence. The
absolute error is calculated on the proportion in the max_months_waited bin
relative to the total waiting (even if a non-percentage target is used)}
}
\value{
a capacity multiplier representing the annual change in capacity
  (from the input t_1_capacity to a capacity at t = 13) to achieve the
  desired target within the target tolerance
}
\description{
Optimise the capacity profile for projections
}
