% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steady_state.R
\name{optimise_steady_state}
\alias{optimise_steady_state}
\title{Identify the steady-state solution with the closest treatment capacity or renege rate
  to a given target}
\usage{
optimise_steady_state(
  referrals,
  target,
  renege_params,
  method = c("treatments", "renege_rates"),
  target_time = 4 + (68/487),
  percentile = 0.92,
  tolerance = target * 0.05,
  max_iterations = 10
)
}
\arguments{
\item{referrals}{numeric of length 1; value representing the steady-state number
of referrals.}

\item{target}{Numeric. The desired treatment capacity/renege rate to be achieved, see
'method' argument to define which one will be used.}

\item{renege_params}{numeric; proportion \ifelse{html}{\out{(0 &le; r<sub>m</sub> &leq; 1)}}{\eqn{(0 \leq r_m \leq 1)}}
of waiting patients reneging in the m-th month of waiting, where \ifelse{html}{\out{r<sub>2</sub>}}{\eqn{r_2}} = 0.1
refers to 10\% of those in the second month of waiting reneging.}

\item{method}{String, length 1. Can take value "treatments" (default) to optimise based
on treatment targets, or "renege_rates" to optimise on renege rates}

\item{target_time}{numeric of length 1; the target waiting time,
in months, to achieve the specified percentile.}

\item{percentile}{Numeric value between 0 and 1 indicating the desired percentile (default is 0.92).}

\item{tolerance}{Numeric. Acceptable deviation from `target` for convergence.
Defaults to 10\% of `target`.}

\item{max_iterations}{Integer. Maximum number of iterations for the binary search.
Defaults to 10.}
}
\value{
A list containing:
  \item{p1}{The value of \code{p1} found to achieve the target waiting time.}
  \item{time_p}{The interpolated waiting time at the specified percentile.}
  \item{mu}{The total number of treatments required when the solution is found.}
  \item{wlsize}{The total size of the waiting list when the solution is found.}
  \item{waiting_list}{The data frame with historical data formatted and waiting
    list sizes calculated.}
  \item{niterations}{The number of iterations performed in the binary search.}
  \item{status}{Converged indicates a solution was found, and Not converged
    indicates no solution was found.}
  \item{method}{The method the solution was identified.}
}
\description{
This function performs a binary search to identify a treatment capacity or
  renege rate that solves the steady state criteria closest to a specified target
  treatment capacity `mu` or renege rate. It uses the `find_p()` function to evaluate
  system convergence and iteratively adjusts `mu_1` until the solution is within a
  defined tolerance or the maximum number of iterations is reached. If convergence is
  not achieved, the function performs secondary searches to identify the best available
  steady-state solution.
}
\examples{
\dontrun{
optimise_steady_state(
  referrals = 100,
  target = 80,
  renege_params = runif(24),
  method = "treatments"
)
}

}
