% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tidy_file}
\alias{tidy_file}
\title{Make the excel file a tidy format}
\usage{
tidy_file(excel_filepath, sheet = "Provider", n_skip)
}
\arguments{
\item{sheet}{string; sheet name}

\item{n_skip}{number of rows to skip before reading in main table from sheet}
}
\value{
a tidy tibble
trust, specialty, period, type,
#'   months_waited and value
trust, specialty, period, type,
#'   months_waited and value
}
\description{
Make the excel file a tidy format
}
